+++ src/pages/about/_our-goals.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import styled from 'styled-components'
-import Container from '../../components/containers/container.js'
-import { Card } from '../../components/elements/card.js'
-import { localize } from '../../components/localization'
+import Container from 'components/containers/container.js'
+import { Card } from 'components/elements/card.js'
+import { localize } from 'components/localization'
 
 import VisionSVG from 'images/svg/vision.svg'
 import FuseeSVG from 'images/svg/fusee.svg'
