+++ src/pages/about/_our-locations.js
@@ -1,10 +1,10 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
-import Button from '../../components/form/button.js'
-import { localize } from '../../components/localization'
-import Container from '../../components/containers/container.js'
-import { Header } from '../../components/elements/typography.js'
+import Button from 'components/form/button.js'
+import { localize } from 'components/localization'
+import Container from 'components/containers/container.js'
+import { Header } from 'components/elements/typography.js'
 
 import DubaiSVG from 'images/svg/dubai.svg'
 import KualaLumpurSVG from 'images/svg/kuala-lumpur.svg'
