+++ src/pages/about/index.js
@@ -1,10 +1,10 @@
 import React from 'react'
-import { localize, WithIntl } from '../../components/localization'
-import SEO from '../../components/containers/seo'
-import Layout from '../../components/layout/layout'
-import Hero from '../../components/elements/hero.js'
-import Modal, { useModal } from '../../components/elements/modal'
-import SignupModal from '../../components/elements/signup-modal'
+import { localize, WithIntl } from 'components/localization'
+import SEO from 'components/containers/seo'
+import Layout from 'components/layout/layout'
+import Hero from 'components/elements/hero.js'
+import Modal, { useModal } from 'components/elements/modal'
+import SignupModal from 'components/elements/signup-modal'
 
 import { OurLocations } from './_our-locations'
 import { OurNumbers } from './_our-numbers'
