+++ src/pages/help-centre/_article-section.js
@@ -1,10 +1,10 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
-import { Text, Header } from '../../components/elements/typography'
-import { localize } from '../../components/localization'
-import { StyledLink } from '../../components/elements/link'
-import Container from '../../components/containers/container'
+import { Text, Header } from 'components/elements/typography'
+import { localize } from 'components/localization'
+import { StyledLink } from 'components/elements/link'
+import Container from 'components/containers/container'
 import { ListWithLinks } from './_list'
 import SearchIcon from 'images/svg/search.svg'
