+++ src/pages/help-centre/_search-results.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
-import { Text, Header } from '../../components/elements/typography'
-import { localize } from '../../components/localization'
+import { Text, Header } from 'components/elements/typography'
+import { localize } from 'components/localization'
 import { Ul, Li, ListWithLinks } from './_list'
 
 const ErrorHeader = styled(Header)`
