+++ src/pages/keep-safe/_icon-grid.js
@@ -2,8 +2,8 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
 import device from 'themes/device'
-import { Text } from '../../components/elements/typography.js'
-import { localize } from '../../components/localization'
+import { Text } from 'components/elements/typography.js'
+import { localize } from 'components/localization'
 
 import Two from 'images/svg/two-factor.svg'
 import DifferentPassword from 'images/svg/different-password.svg'
